Reorg mem files
